requirements.txt
pillow
Flask
numpy
tensorflow
scikit-image
face_recognition @ git+https://github.com/thetoby9944/face_recognition